==> notebooks/01c_data_checks.ipynb <==
## Spanish injection

import pandas as pd
from pathlib import Path
import os


df_3k_neg_long_esp = pd.read_csv(Path(os.getcwd()).parent / "data/negative_3k_es_prompts.csv")
df_700_neg_short_esp = pd.read_csv(Path(os.getcwd()).parent / "data/negative_700_es_prompts.csv")

df_700_neg_short_esp[df_700_neg_short_esp["tipo"].isna()]

df_700_neg_short_esp["annotation_1"] = "injection"
df_700_neg_short_esp = df_700_neg_short_esp.rename({"tipo":"annotation_2"}, axis=1)
df_700_neg_short_esp["lang"] = "spanish"
df_700_neg_short_esp["label"] = "1"

df_700_neg_short_esp.head(3)

df_700_neg_short_esp["annotation_2"].value_counts()

df_700_neg_short_esp["annotation_2"]=df_700_neg_short_esp["annotation_2"].map(
    {
        "Inyección Directa": "direct_injection", 
        "Inyección Indirecta": "indirect_injection",
        "Roleplay Jailbreak": "roleplay_jailbreak",
        "Token Smuggling": "token_smuggling",
        "Inyección de Código": "code_injection",
        "Inyección Contextual": "contextual_injection",
        "Reformulación Creativa": "creative_rewriting",
        "Inyección de Código": "code_injection",
     }
)

df_700_neg_short_esp.head(3)

df_3k_neg_long_esp = df_3k_neg_long_esp.rename({"tipo": "label"}, axis=1)
df_3k_neg_long_esp["lang"] = "spanish"
df_3k_neg_long_esp["annotation_1"] = "injection"
df_3k_neg_long_esp["annotation_2"] = "long_format_prompt"
df_3k_neg_long_esp["label"] = df_3k_neg_long_esp["label"].astype("object")

df_3k_neg_long_esp.head(3)

df_neg_concat_spanish = pd.concat(
    [df_3k_neg_long_esp, df_700_neg_short_esp]
)
df_neg_concat_spanish.to_csv(Path(os.getcwd()).parent / "data/processed/df_injection_spanish.csv", index=False)

## Spanish benign

df_benign_2k_es_prompts = pd.read_csv(Path(os.getcwd()).parent / "data/benign_2k_es_prompts.csv")

df_benign_2k_es_prompts = df_benign_2k_es_prompts.rename({"tipo": "label"}, axis=1)
df_benign_2k_es_prompts["lang"] = "spanish"
df_benign_2k_es_prompts["annotation_1"] = "benign"
df_benign_2k_es_prompts["annotation_2"] = ""
df_benign_2k_es_prompts["label"] = df_benign_2k_es_prompts["label"].astype("object")

df_benign_2k_es_prompts.to_csv(Path(os.getcwd()).parent / "data/processed/df_benign_spanish.csv", index=False)

df_spanish_concat = pd.concat(
    [df_neg_concat_spanish, df_benign_2k_es_prompts]
)
df_spanish_concat.head(3)

df_spanish_concat.to_csv(Path(os.getcwd()).parent / "data/merged/df_spanish.csv", index=False)

## English injection
